Add an updated format for task step specification
The new method is tied into the HMM work. This is likely globally not
ideal, but the sepration is not understood at this time.
When understood, we may separate task configuration with HMM
configuration.
This new file would presumably be the new configuration for the next
generation of the task monitor.