Add ROS_LOCALHOST_ONLY variable specification
Experimented a little with FastDDS and CycloneDDS configuration in an
attempt to observe which could be firmly configured to not use the
network stack or just loopback. It seems that FastRTPS should be, and
is, using shared-memory to communicate since everything is on the same
machine. Adding a ROS_LOCALHOST_ONLY=1 environment variable should
prevent any outside communication, in or out. It is still unknown why
FastRTPS would stall out a node given local shared-memory communication,
but this has not been observed yet empirically since adding the
ROS_LOCALHOST_ONLY variable setting.