Fix misaligned configuration path
GSP had a hard-coded path as a defaulted constructor parameter.
This value is incorrect given some reorganization if config files that
must have happened.
Offending nodes now take in an explicit parameter to reference the
relevant configuration file and now explicitly passes that to the GSP
instance that is being constructed.