Simplify docker images
We have yet to take make use of any layered docker image semantics due
to the establishment of the current use-case. Thus, removed the
complexity that was having the base "dev" image to prefer to have just
one docker file for the whole environment. We can reintroduce a "dev"
vs. "runtime" base layers at a later time if it becomes pertinent.