Update docker build script to ignore local-only build artifact
Node modules "installed" are left on the host via mounted directories.
This would cause an annoyance where to build new docker, clean docker
images, this directory would have to be removed every time. This change
adds this fixed, known directory to the dockerignore file, so it will
never be considered for addition to a docker image build, and the build
script was updated to appropriately ignore this file in the clean check.